[RENAME] 타입스크립트 절대경로 변환
