+++ commonUi/Icons/index.tsx
@@ -159,7 +159,7 @@ const SearchIcon40 = () => (
   </svg>
 );
 
-const CloseMenuIcon = ({ color = '#595959' }: { color: string }) => (
+const CloseMenuIcon = ({ color = '#595959' }: { color?: string }) => (
   <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
     <path
       fillRule="evenodd"
