+++ commonUi/Tags/index.ts
@@ -1,0 +1,2 @@
+export * from './ApplyTag';
+export * from './CompanyAgeTag';
