+++ components/CTASection/index.ts
@@ -1,0 +1,1 @@
+export { default as CTASection } from './CTASection';
