+++ components/FeaturedSupportPrograms/TabletMobileSlideCards/TabletMobileSlideCards.styled.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { MediaQuery, FontSize, FontWeight } from '../../../utils';
+import { MediaQuery, FontSize, FontWeight } from 'utils';
 
 export const SlideCardWrapper = styled.div`
   display: none;
