+++ components/Footer/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Footer';
+export { default as Footer } from './Footer';
