+++ components/Header/index.ts
@@ -1,3 +1,1 @@
-import Header from './Header';
-
-export { Header };
+export { default as Header } from './Header';
