--- components/ResponsiveSection/ResponsiveSection.styled.tsx
@@ -1,2 +1,0 @@
-import styled from '@emotion/styled';
-import { MediaQuery, FontSize, FontWeight } from '../../utils';
