--- components/ResponsiveSection/ResponsiveSection.tsx
@@ -1,5 +1,0 @@
-function ResponsiveSection(mediaQuery) {
-  const test = window.matchMedia(mediaQuery).matches;
-}
-
-export default ResponsiveSection;
