--- components/ResponsiveSection/index.ts
@@ -1,1 +1,0 @@
-import { default as ResponsiveSection } from './ResponsiveSection';
