+++ components/SupportProgramBanners/SupportProgramBanners.styled.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { MediaQuery, FontSize, FontWeight } from '../../utils';
+import { MediaQuery, FontSize, FontWeight } from 'utils';
 
 export const Contatiner = styled.section`
   margin-bottom: 60px;
