+++ components/SupportPrograms/SupportProgramFilters/HostSearch/HostSearch.styled.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { FontSize, FontWeight } from '../../../../utils';
+import { FontSize, FontWeight } from 'utils';
 
 export const Wrapper = styled.article`
   position: relative;
