+++ components/SupportPrograms/SupportProgramFilters/HostSearch/HostSearch.tsx
@@ -1,7 +1,6 @@
 import { useState } from 'react';
-
+import Search24 from 'commonUi/Icons/Search24/search24.svg';
 import * as Styled from './HostSearch.styled';
-import Search24 from '../../../../commonUi/Icons/Search24/search24.svg';
 
 import { contain } from '../../SupportPrograms.utils';
