+++ components/SupportPrograms/SupportProgramFilters/TypeFilters/TypeFilters.styled.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { FontSize, FontWeight, MediaQuery } from '../../../../utils';
+import { FontSize, FontWeight, MediaQuery } from 'utils';
 
 export const TypeFilterList = styled.ul`
   display: flex;
