+++ graphql/index.ts
@@ -1,2 +1,4 @@
 export * from './types';
 export * from './utils';
+export * from './queries';
+export { default as client } from './client';
